+++ src/Evaluation/ScanEvaluationBase.cpp
@@ -344,7 +344,7 @@ CEvaluationBase* ScanEvaluationBase::FindOptimumShiftAndSqueezeFromFraunhoferRef
         CFitWindow improvedFitWindow = fitWindow;
 
         // The fit is good enough to use the values
-        for (int it = 0; it < improvedFitWindow.nRef; ++it)
+        for (size_t it = 0; it < improvedFitWindow.nRef; ++it)
         {
             improvedFitWindow.ref[it].m_shiftOption = SHIFT_TYPE::SHIFT_FIX;
             improvedFitWindow.ref[it].m_squeezeOption = SHIFT_TYPE::SHIFT_FIX;
